Split install-kiosk-scripts into update and no-update script
This helps a lot with testing since the no-update script calls other
kiosk-scripts commands with respect to the local folder, not the global
installation folder, which is supposed to be in sync with the git repo.

Local changes to thge kiosk scripts can be tested with commiting to git
and pushing them to the online repo.