Read certain scripts completely before executing long-lasting commands
When piping scripts from curl into a shell, commands are executed line
by line. If a command takes a long time to complete (such as apt-get),
a connection timeout might occur such that the remaining script lines
will not be executed. Wrapping all commands into a function and only
calling that function at the end of the script makes sure the
entire script has been downloaded before execution is started.