Add delays to work around race conditions in the kiosk-browser
Only one instance of the kiosk-browser should run at a time. Complex
exhibits like `con-espressione` have multiple browser windows which may
cause the kiosk-browser to crash. Adding some delays makes it a lot less
likely to happen, but it should be made absolutely clear that this
should be considered a temperorary workaround, not a proper solution.