Reduce `ai-jam` audio buffer size to 128 samples
By default, the Chromium process that serves as the basis of the
kiosk-browser uses a buffer size that seems to add about 100ms of audio
latency, which is way to much for playing virtual instruments.

This patch reduces the latency to about 3ms (which may be too low for
some applications and result in dropouts). Additional latency may be
added by the audio system of the OS.