Always rebind all joysticks in deterministic order
Needing to specify the sysfs path to the joysticks that should be
rebound is cumbersome and doesn't work well if the same configuration is
moved to different hardware, sometimes even different version of the OS.
Therefore, the explicit device selection has been abandoned in favor of
an always deterministic assignment of device ids and devices given the
same hardware and system configuration.

The joystick order can be manipulated externally by swapping the ports
the devices are connected to.